--- src/platform/plugins/shared/dashboard/server/api/transforms/in/transform_panel_in_error.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-export class TransformPanelInError extends Error {
-  public readonly type: string;
-  public readonly config: object;
-
-  constructor(message: string, type: string, config: object) {
-    super(message);
-    this.name = 'TransformPanelInError';
-    this.type = type;
-    this.config = config;
-  }
-}
